# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuthSys API",
    "version": "23.0.0",
    "description": "Private OAuth 2.0 and OpenID Connect provider. Authorization Code + PKCE is the supported browser flow."
  },
  "servers": [
    { "url": "https://testmono-user.duckdns.org", "description": "Current HTTPS deployment" },
    { "url": "http://localhost:8080", "description": "Local Docker fallback" }
  ],
...
  "paths": {
...
    "/api/oauth/authorize": {
      "get": {
        "tags": ["OAuth/OIDC"],
        "summary": "Start Authorization Code flow",
        "parameters": [
          { "name": "response_type", "in": "query", "required": true, "schema": { "type": "string", "enum": ["code"] } },
          { "name": "client_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "redirect_uri", "in": "query", "required": true, "schema": { "type": "string", "format": "uri" } },
          { "name": "scope", "in": "query", "required": true, "schema": { "type": "string", "example": "openid profile email" } },
          { "name": "state", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "nonce", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "code_challenge", "in": "query", "required": true, "schema": { "type": "string", "minLength": 43, "maxLength": 128 } },
          { "name": "code_challenge_method", "in": "query", "required": true, "schema": { "type": "string", "enum": ["S256"] } },
          { "name": "prompt", "in": "query", "schema": { "type": "string", "enum": ["none"] } }
        ],
        "responses": {
          "302": { "description": "Login/consent page or registered redirect URI" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "post": {
        "tags": ["OAuth/OIDC"],
        "summary": "Allow or deny the displayed consent request",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "JSON redirect_url containing a code or OAuth error" } }
      }
    },
...
    "/api/oauth/clients": {
      "get": {
        "tags": ["Clients"],
        "summary": "List OAuth clients owned by the current user",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated clients without client secrets" } }
      },
      "post": {
        "tags": ["Clients"],
        "summary": "Register a confidential web client",
        "description": "The plaintext client_secret is shown only once.",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClientRegistration" } } }
        },
        "responses": {
          "201": { "description": "Client ID and one-time client secret" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/oauth/clients/{clientId}": {
      "get": {
        "tags": ["Clients"],
        "summary": "Read an owned client",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClientId" }],
        "responses": { "200": { "description": "Client metadata" }, "404": { "description": "Client not found" } }
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Update mutable client metadata",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClientId" }],
        "responses": { "200": { "description": "Updated client" }, "404": { "description": "Client not found" } }
      },
      "delete": {
        "tags": ["Clients"],
        "summary": "Deactivate an owned client and revoke its consents",
        "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ClientId" }],
        "responses": { "200": { "description": "Client deactivated" }, "404": { "description": "Client not found" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": { "type": "apiKey", "in": "cookie", "name": "connect.sid" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "clientBasic": { "type": "http", "scheme": "basic", "description": "OAuth client_id and client_secret" }
    },
    "parameters": {
      "ClientId": { "name": "clientId", "in": "path", "required": true, "schema": { "type": "string", "maxLength": 128 } }
    },
    "schemas": {
...
      "ClientRegistration": {
        "type": "object",
        "required": ["client_name", "redirect_uris"],
        "properties": {
          "client_name": { "type": "string", "minLength": 2, "maxLength": 100 },
          "redirect_uris": { "type": "array", "minItems": 1, "maxItems": 10, "items": { "type": "string", "format": "uri" } },
          "description": { "type": "string", "maxLength": 500 },
          "logo_uri": { "type": "string", "format": "uri" },
          "contact_email": { "type": "string", "format": "email" }
        }
      }
    },
    "responses": {
      "BadRequest": { "description": "Invalid or unsupported request" },
...
    }
  }
}